Move testing module to deno_std
Upgrades deno_std submodule.
